+++ src/test/java/org/stellar/sdk/SorobanServerTest.java
@@ -661,6 +661,7 @@ public class SorobanServerTest {
         resp.getEvents().get(0).getContractId(),
         "CBQHNAXSI55GX2GN6D67GK7BHVPSLJUGZQEU7WJ5LKR5PNUCGLIMAO4K");
     assertEquals(resp.getEvents().get(0).getId(), "0000000459561504768-0000000000");
+    assertEquals(resp.getEvents().get(0).getPagingToken(), "0000000459561504768-0000000000");
     assertEquals(resp.getEvents().get(0).getTopic().size(), 2);
     assertEquals(resp.getEvents().get(0).getTopic().get(0), "AAAADwAAAAdDT1VOVEVSAA==");
     assertEquals(resp.getEvents().get(0).getTopic().get(1), "AAAADwAAAAlpbmNyZW1lbnQAAAA=");